student API

+++ StudentData.Domain.Core/StudentGroup.cs
@@ -6,10 +6,10 @@ namespace StudentData.Domain.Core
 {
     public class StudentGroup
     {
-        public int StudentId { get; set; }
+        public Int64 StudentId { get; set; }
         public Student Student { get; set; }
 
-        public int GroupId { get; set; }
+        public Int64 GroupId { get; set; }
         public Group Group { get; set; }
     }
 }
